short type fix / duplicated item warning

+++ voxels-common/src/main/java/common/game/Item.java
@@ -2,17 +2,17 @@ package common.game;
 
 public class Item {
 
-  private final int id;
+  private final short id;
   private final String name;
   private final int maxStackSize;
 
-  public Item(int id, String name, int maxStackSize) {
-    this.id = (short) id;
+  public Item(short id, String name, int maxStackSize) {
+    this.id = id;
     this.name = name;
     this.maxStackSize = maxStackSize;
   }
 
-  public int getId() {
+  public short getId() {
     return id;
   }
